+++ package.json
@@ -13,7 +13,8 @@
     "mailserver": "node ./src/backend/mailserver.js",
     "predeploy": "vue-cli-service build --mode prod",
     "deploy": "gh-pages -d dist",
-    "clean": "gh-pages-clean"
+    "clean": "gh-pages-clean",
+    "test": "jest --detectOpenHandles --silent=false --forceExit"
   },
   "dependencies": {
     "@fullcalendar/core": "^5.10.1",
@@ -30,10 +31,12 @@
     "dotenv": "^16.0.0",
     "express": "^4.17.3",
     "express-validator": "^6.14.0",
+    "jest": "^27.5.1",
     "node-cron": "^3.0.0",
     "nodemailer": "^6.7.2",
     "nodemailer-smtp-transport": "^2.7.4",
     "pg": "^8.7.3",
+    "supertest": "^6.2.2",
     "v-calendar": "^3.0.0-alpha.8",
     "vue": "^3.2.13",
     "vue-cli": "^2.9.6",
